+++ app/src/module/spark/normalization/__init__.py
@@ -1,0 +1,1 @@
+from .helpers import minmax_normalization