+++ app/src/module/spark/normalization/constants.py
@@ -1,0 +1,7 @@
+CONSTANTS = {
+    'COLS': {
+        'VECTORIZED_COL': 'x_vec',
+        'SCALED_COL': 'x_scaled',
+    }
+}
+"""The normalization constants"""